+++ urbansim_templates/data/__init__.py
@@ -1,1 +1,2 @@
 from .io import TableFromDisk
+from .save_data import SaveData
