+++ urbansim_templates/data/io.py
@@ -31,7 +31,7 @@ class TableFromDisk():
     be able to use it as a join key. Following these naming conventions eliminates the 
     need for Orca "broadcasts".
     
-    All the parameters can also be set as properties of the class instance.
+    All the parameters can also be set as properties after creating the class instance.
     
     Parameters
     ----------
@@ -308,161 +308,3 @@ class TableFromDisk():
         return True
     
     
-@modelmanager.template
-class SaveData():
-    """
-    Class for saving Orca tables to local CSV or HDF5 files.
-    
-    All the parameters can also be set as properties of the class instance.
-    
-    Parameters
-    ----------
-    output_type : 'csv' or 'hdf', optional
-        This is required to save the table, but does not have to be provided when the 
-        object is created.
-    
-    path : str, optional
-        Local file path to save the data to, either absolute or relative to the 
-        ModelManager config directory. The string you provide will immediately be 
-        normalized to a platform-agnostic format, using `os.path.normpath()` in Python 2 
-        or `pathlib.Path()` in Python 3. It is always safe to provide a Unix-style path, 
-        and you may provide a Windows-style path if you are creating the model step in 
-        Windows. Saved steps will run on any platform. 
-    
-    extra_settings : dict, optional
-        Additional arguments to pass to `pd.to_csv()` or `pd.to_hdf()`. For example, you
-        could automatically compress csv data using {'compression': 'gzip'}, or specify 
-        a custom table name for an hdf store using {'key': 'table-name'}. See Pandas 
-        documentation for additional settings.
-            
-    name : str, optional
-        Name of the Orca table. HOW TO NAME THE MODEL STEP?
-    
-    tags : list of str, optional
-        Tags, passed to ModelManager.
-    
-    """
-    def __init__(self, 
-            output_type = None, 
-            path = None, 
-            extra_settings = {}, 
-            name = None,
-            tags = [], 
-        
-        # Template-specific params
-        self.output_type = output_type
-        self.path = path
-        self.extra_settings = extra_settings
-        
-        # Standard params
-        self.name = name
-        self.tags = tags
-        
-        # Automatic params
-        self.template = self.__class__.__name__
-        self.template_version = __version__
-    
-    
-    @classmethod
-    def from_dict(cls, d):
-        """
-        Create an object instance from a saved dictionary representation.
-        
-        Parameters
-        ----------
-        d : dict
-        
-        Returns
-        -------
-        Table
-        
-        """
-        obj = cls(
-            output_type = d['output_type'],
-            path = d['path'],
-            extra_settings = d['extra_settings'],
-            name = d['name'],
-            tags = d['tags'],
-        )
-        return obj
-    
-    
-    def to_dict(self):
-        """
-        Create a dictionary representation of the object.
-        
-        Returns
-        -------
-        dict
-        
-        """
-        d = {
-            'template': self.template,
-            'template_version': self.template_version,
-            'name': self.name,
-            'tags': self.tags,
-            'output_type': self.output_type,
-            'path': self.path,
-            'extra_settings': self.extra_settings,
-        }
-        return d
-    
-    
-    @property
-    def path(self):
-        return self.__path
-    @path.setter
-    def path(self, value):
-        if value is not None:
-            try:
-                value = str(pathlib.Path(value))  # Python 3.4+
-            except:
-                value = os.path.normpath(value)
-        self.__path = value
-            
-
-    def run(self):
-        """
-        Register a data table with Orca.
-        
-        Requires values to be set for ``source_type``, ``name``, and ``path``. CSV data 
-        also requires ``csv_index_cols``. 
-        
-        Returns
-        -------
-        None
-        
-        """
-        if self.source_type not in ['csv', 'hdf']:
-            raise ValueError("Please provide a source type of 'csv' or 'hdf'")
-        
-        if self.name is None:
-            raise ValueError("Please provide a table name")
-        
-        if self.path is None:
-            raise ValueError("Please provide a file path")
-        
-        kwargs = self.extra_settings
-        
-        # Table from CSV file
-        if self.source_type == 'csv':
-            if self.csv_index_cols is None:
-                raise ValueError("Please provide index column name(s) for the csv")
-        
-            @orca.table(table_name = self.name, 
-                        cache = self.cache, 
-                        cache_scope = self.cache_scope, 
-                        copy_col = self.copy_col)
-            def orca_table():
-                df = pd.read_csv(self.path, **kwargs).set_index(self.csv_index_cols)
-                return df
-            
-        # Table from HDF file
-        elif self.source_type == 'hdf':
-            @orca.table(table_name = self.name, 
-                        cache = self.cache, 
-                        cache_scope = self.cache_scope, 
-                        copy_col = self.copy_col)
-            def orca_table():
-                df = pd.read_hdf(self.path, **kwargs)
-                return df
